% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validation.R
\name{check_probabilities_for_range}
\alias{check_probabilities_for_range}
\title{This function checks if the survival probabilities are in the range of 0 to 1.}
\usage{
check_probabilities_for_range(prob)
}
\arguments{
\item{prob}{A vector with values representing the survival probabilities.}
}
\description{
This function checks if the survival probabilities are in the range of 0 to 1.
}
