% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validation.R
\name{check_risk_information}
\alias{check_risk_information}
\title{This function does some basic checks on the time of risk and number of risk information.}
\usage{
check_risk_information(time_of_risk, number_at_risk)
}
\arguments{
\item{time_of_risk}{A vector with time points at which patient numbers at risk are provided.}

\item{number_at_risk}{A vector with number of patients at risk.}
}
\description{
This function does some basic checks on the time of risk and number of risk information.
}
