% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validation.R
\name{check_valid_numbers}
\alias{check_valid_numbers}
\title{This function does some basic checks for numeric input.}
\usage{
check_valid_numbers(input)
}
\arguments{
\item{input}{A vector with different values which are supposed to be numeric.}
}
\description{
This function does some basic checks for numeric input.
}
