% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validation.R
\name{find_non_decreasing_probabilities}
\alias{find_non_decreasing_probabilities}
\title{This function checks if there are any non decreasing survival probabilities.
It sorts by survival times in ascending order first and then checks if the survival probabilities are descending.}
\usage{
find_non_decreasing_probabilities(time, prob)
}
\arguments{
\item{time}{A vector with survival time.}

\item{prob}{A vector with survival probabilities.}
}
\description{
This function checks if there are any non decreasing survival probabilities.
It sorts by survival times in ascending order first and then checks if the survival probabilities are descending.
}
