% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validation.R
\name{find_non_increasing_survival_times}
\alias{find_non_increasing_survival_times}
\title{This function checks if there are any non increasing survival time.
It sorts by survival probabilities in descending order first and then checks if the survival time are ascending.}
\usage{
find_non_increasing_survival_times(time, prob)
}
\arguments{
\item{time}{A vector with survival time.}

\item{prob}{A vector with survival probabilities.}
}
\description{
This function checks if there are any non increasing survival time.
It sorts by survival probabilities in descending order first and then checks if the survival time are ascending.
}
