% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ipd.R
\name{fix_non_decreasing_survival_probabilities}
\alias{fix_non_decreasing_survival_probabilities}
\title{This function sorts first the input by survival times in ascending order,
then fixes survival probabilities by replacing non-decreasing ones with predecessor values.}
\usage{
fix_non_decreasing_survival_probabilities(time, prob)
}
\arguments{
\item{time}{A vector of survival times.}

\item{prob}{A vector of survival probabilities.}
}
\value{
A data.frame of modified survival probabilities and times.
}
\description{
This function sorts first the input by survival times in ascending order,
then fixes survival probabilities by replacing non-decreasing ones with predecessor values.
}
