% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ipd.R
\name{fix_non_increasing_survival_times}
\alias{fix_non_increasing_survival_times}
\title{This function sorts first the input by survival probabilities in descending order,
then fixes survival times by replacing non-increasing ones with predecessor values.}
\usage{
fix_non_increasing_survival_times(time, prob)
}
\arguments{
\item{time}{A vector of survival times.}

\item{prob}{A vector of survival probabilities.}
}
\value{
A data.frame of modified survival times and probabilities.
}
\description{
This function sorts first the input by survival probabilities in descending order,
then fixes survival times by replacing non-increasing ones with predecessor values.
}
