% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ipd.R
\name{get_lower_indices}
\alias{get_lower_indices}
\title{This function uses a for loop to find the minimum index of elements in the \code{time} vector,
which are the survival times, that are greater than or equal to each element in the \code{trisk} vector,
which are the temporal intervals.}
\usage{
get_lower_indices(time, trisk)
}
\arguments{
\item{time}{A vector of survival probabilities.}

\item{trisk}{A vector of times at which patients are at risk.}
}
\value{
A vector that contains the lower time limits for each interval.
It defines the starting point of each interval where calculations for survival
estimates and events will be made.
}
\description{
This function uses a for loop to find the minimum index of elements in the \code{time} vector,
which are the survival times, that are greater than or equal to each element in the \code{trisk} vector,
which are the temporal intervals.
}
