% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ipd.R
\name{get_upper_indices}
\alias{get_upper_indices}
\title{This function uses a for loop to find the maximum index of elements in the \code{time} vector,
which are the survival times, that are less than or equal to each element in the \code{trisk} vector,
which are the temporal intervals.}
\usage{
get_upper_indices(time, trisk)
}
\arguments{
\item{time}{A vector of survival probabilities.}

\item{trisk}{A vector of times at which patients are at risk.}
}
\value{
A vector that contains the upper time limits for each interval.
It defines the end point of each interval where calculations will stop.
}
\description{
This function uses a for loop to find the maximum index of elements in the \code{time} vector,
which are the survival times, that are less than or equal to each element in the \code{trisk} vector,
which are the temporal intervals.
}
