% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_point_translation.R
\name{pixel_to_data_point}
\alias{pixel_to_data_point}
\title{This function uses linear interpolation to translate from on screen pixel space into data space given reference values.}
\usage{
pixel_to_data_point(pixel_x, pixel_y, x_ref, y_ref)
}
\arguments{
\item{pixel_x}{A pixel value in x dimension.}

\item{pixel_y}{A pixel value in y dimension.}

\item{x_ref}{A 2x2 matrix which contains the reference values in x dimension in data and pixel space.}

\item{y_ref}{A 2x2 matrix which contains the reference values in y dimension in data and pixel space.}
}
\value{
A vector with the data values inx and y dimension
}
\description{
This function uses linear interpolation to translate from on screen pixel space into data space given reference values.
}
