% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ipd.R
\name{prepare_time_interval_data}
\alias{prepare_time_interval_data}
\title{This function provides a data.frame with all the relevant time interval information.}
\usage{
prepare_time_interval_data(time, trisk, nrisk, lower, upper)
}
\arguments{
\item{time}{A vector of survival probabilities.}

\item{trisk}{A vector of times at which patients are at risk.}

\item{nrisk}{A vector of the number of patients at risk.}

\item{lower}{A vector that contains the lower time limits for each interval.
It defines the starting point of each interval where calculations for survival estimates and events will be made.}

\item{upper}{A The \code{upper} parameter is also a vector that contains the upper time limits for each interval.
It defines the end point of each interval where calculations will stop.}
}
\value{
A data.frame with all the relevant time interval information.
}
\description{
This function provides a data.frame with all the relevant time interval information.
}
