% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validation.R
\name{validation_input_guyot}
\alias{validation_input_guyot}
\title{This function does some basic validations of the input parameters for reconstructing
IPD data by the method proposed by Guyot et al.}
\usage{
validation_input_guyot(time, prob, trisk, nrisk, lower, upper, tot_events = NA)
}
\arguments{
\item{time}{A vector of the x coordinates of the survival times.}

\item{prob}{A vector of the y coordinates corresponding to 'time', between 0 and 1.
These represent survival probabilities at the times in 'time'.}

\item{trisk}{A vector of times at which patients are at risk.}

\item{nrisk}{A vector of the number of patients at risk.}

\item{lower}{A vector that contains the lower time limits for each interval.
It defines the starting point of each interval where calculations for survival estimates and events will be made.}

\item{upper}{A The \code{upper} parameter is also a vector that contains the upper time limits for each interval.
It defines the end point of each interval where calculations will stop.}

\item{tot_events}{The number total events (optional).}
}
\description{
This function does some basic validations of the input parameters for reconstructing
IPD data by the method proposed by Guyot et al.
}
