% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_validation.R
\name{validation_input_rogula}
\alias{validation_input_rogula}
\title{This function does some basic validations of the input parameters for reconstructing
IPD data by the method proposed by Rogula et al.}
\usage{
validation_input_rogula(n, time, prob, cens_t = NA)
}
\arguments{
\item{n}{The number at risk at time zero.}

\item{time}{A vector of the x coordinates of the bottoms of the drops in survival.}

\item{prob}{A vector of the y coordinates corresponding to 'time', between 0 and 1.
These represent survival proportions at the times in t.}

\item{cens_t}{A vector of times at which patients were censored (optional).}
}
\description{
This function does some basic validations of the input parameters for reconstructing
IPD data by the method proposed by Rogula et al.
}
